requirements.txt
numpy
matplotlib
scipy
scikit-image
hickle
Pillow
jieba


